api/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.4
